+++ mdcx/config/v1.py
@@ -121,7 +121,7 @@ class ConfigV1:
     website_single: str = r"airav_cc"
     website_youma: str = r"airav_cc,iqqtv,javbus,freejavbt,dmm,javlibrary,javdb,avsex,avmoo,lulubar,xcity,avsox,javfree"
 
-    website_wuma: str = r"iqqtv,javbus,freejavbt,avsox,javdb"
+    website_wuma: str = r"iqqtv,javbus,freejavbt,avsox,javdb,aventertainments"
     website_suren: str = r"mgstage,avsex,freejavbt,javbus,javdb"
     website_fc2: str = r"fc2,freejavbt,javdb,avsox,javfree"
     website_oumei: str = r"theporndb,avheat,javdb,javbus"